+++ src/common/helpers/events.ts
@@ -1,0 +1,7 @@
+export const debounce = (fn: Function, ms = 100) => {
+  let timer: ReturnType<typeof setTimeout>;
+  return function (e: Event) {
+    clearTimeout(timer);
+    timer = setTimeout(() => fn.apply(e), ms);
+  };
+};
