+++ src/common/helpers/storybook.ts
@@ -1,0 +1,19 @@
+export const stringToReactHtml = (string: String) => {
+  return { __html: string };
+};
+
+/**
+ * Convert an object to an array of only its values.
+ * Used when importing enums in component stories for populating argType dropdowns.
+ * @param {*} options
+ *  imported enums object
+ */
+export function createOptionsArray(options: object = {}) {
+  const optionsArray: Array<string> = [];
+
+  Object.keys(options).map((key) => {
+    optionsArray.push(options[key as keyof typeof options]);
+  });
+
+  return optionsArray;
+}
