+++ src/components/icon/Icon.mdx
@@ -1,6 +1,6 @@
 import { Meta, Story, Canvas, Controls } from '@storybook/blocks';
 import * as IconStories from './Icon.stories.js';
-import { stringToReactHtml } from '../../common/helpers/helpers.ts';
+import { stringToReactHtml } from '../../common/helpers/storybook';
 
 import logoutIcon from '../../assets/svg/logout.svg';
 import kyndrylIcon from '../../assets/svg/kyndryl-logo.svg';
